requirements.txt
# Python dependencies for Cats vs Dogs Classification Project

# Core ML frameworks
tensorflow>=2.16.0
numpy<2.0,>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0

# Image processing
Pillow==10.0.0
opencv-python==4.8.0.74

# Web framework
fastapi==0.103.0
uvicorn[standard]==0.23.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Jupyter notebook
jupyter==1.0.0
ipykernel==6.25.1
ipywidgets==8.1.0
notebook==7.0.2

# Load testing
locust==2.15.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
